+++ src/checks.jl
@@ -33,12 +33,12 @@ end
 
 
 """
-    check_data(case, modeltype)
+    check_data(case, modeltype, check_timeprofiles::Bool)
 
 Check if the case data is consistent. Use the `@assert_or_log` macro when testing.
 Currently only checking node data.
 """
-function check_data(case, modeltype::EnergyModel)
+function check_data(case, modeltype::EnergyModel, check_timeprofiles::Bool)
     # TODO would it be useful to create an actual type for case, instead of using a Dict with
     # naming conventions? Could be implemented as a mutable in energymodelsbase.jl maybe?
 
@@ -47,24 +47,47 @@ function check_data(case, modeltype::EnergyModel)
     global logs = []
     log_by_element = Dict()
 
+
+    if !check_timeprofiles
+        @warn "Checking of the time profiles is deactivated:\n" *
+        "Deactivating the checks for the time profiles is strongly discouraged. " *
+        "While the model will still run, unexpected results can occur, as well as " *
+        "inconsistent case data.\n\n" *
+        "Deactivating the checks for the timeprofiles should only be considered, " *
+        "when testing new components. In all other instances, it is recommended to " *
+        "provide the correct timeprofiles using a preprocessing routine.\n\n" *
+        "If timeprofiles are not checked, inconsistencies can occur."  maxlog=1
+    end
+
+    # Check the case data. If the case data is not in the correct format, the overall check
+    # is cancelled as extractions would not be possible
+    check_case_data(case)
+    log_by_element["Case data"] = logs
+    if ASSERTS_AS_LOG
+        compile_logs(case, log_by_element)
+    end
+
     𝒯 = case[:T]
 
     for n ∈ case[:nodes]
 
         # Empty the logs list before each check.
         global logs = []
-        check_node(n, 𝒯, modeltype)
+        check_node(n, 𝒯, modeltype, check_timeprofiles::Bool)
         for data ∈ node_data(n)
-            check_node_data(n, data, 𝒯, modeltype)
+            check_node_data(n, data, 𝒯, modeltype, check_timeprofiles)
         end
-        check_time_structure(n, 𝒯)
+
+        if check_timeprofiles
+            check_time_structure(n, 𝒯)
+        end
         # Put all log messages that emerged during the check, in a dictionary with the node as key.
         log_by_element[n] = logs
     end
 
     logs = []
-    check_model(case, modeltype)
-    log_by_element[modeltype] = logs
+    check_model(case, modeltype, check_timeprofiles)
+    log_by_element["Modeltype"] = logs
 
     if ASSERTS_AS_LOG
         compile_logs(case, log_by_element)
@@ -108,16 +131,109 @@ function compile_logs(case, log_by_element)
     end
 end
 
+"""
+    check_case_data(case)
 
-function check_model(case, modeltype::EnergyModel)
-    for p ∈ case[:products]
-        if isa(p, ResourceEmit)
-            @assert_or_log haskey(modeltype.emission_limit, p) "All ResourceEmits requires " *
-                "an entry in the dictionary GlobalData.Emission_limit. For $p there is none."
+Checks the `case` dictionary is in the correct format.
+
+## Checks
+- The dictionary requires the keys `:T`, `:nodes`, `:links`, and `:products`.
+- The individual keys are of the correct type, that is
+  - `:T::TimeStructure`,
+  - `:nodes::Vector{<:Node}`,
+  - `:links::Vector{<:Link}`, and
+  - `:products::Vector{<:Resource}`.
+"""
+function check_case_data(case)
+
+    case_keys = [:T, :nodes, :links, :products]
+    key_map = Dict(
+        :T => TimeStructure,
+        :nodes => Vector{<:Node},
+        :links => Vector{<:Link},
+        :products => Vector{<:Resource},
+    )
+    for key ∈ case_keys
+        @assert_or_log(
+            haskey(case, key),
+            "The `case` dictionary requires the key `:" * string(key) * "` which is " *
+            "not included."
+        )
+        if haskey(case, key)
+            @assert_or_log(
+                isa(case[key], key_map[key]),
+                "The key `" * string(key) * "` in the `case` dictionary contains " *
+                "other types than the allowed."
+            )
         end
     end
 end
 
+"""
+    check_model(case, modeltype::EnergyModel, check_timeprofiles::Bool)
+
+Checks the `modeltype` .
+
+## Checks
+- All `ResourceEmit`s require a corresponding value in the field `emission_limit`.
+- The `emission_limit` time profiles cannot have a finer granulation than `StrategicProfile`.
+- The `emission_price` time profiles cannot have a finer granulation than `StrategicProfile`.
+
+## Conditional checks (if `check_timeprofiles=true`)
+- The profiles in `emission_limit` have to have the same length as the number of strategic
+  periods.
+- The profiles in `emission_price` have to have the same length as the number of strategic
+periods.
+"""
+function check_model(case, modeltype::EnergyModel, check_timeprofiles::Bool)
+
+    𝒯ᴵⁿᵛ = strategic_periods(case[:T])
+
+    # Check for inclusion of all emission resources
+    for p ∈ case[:products]
+        if isa(p, ResourceEmit)
+            @assert_or_log(
+                haskey(emission_limit(modeltype), p),
+                "All `ResourceEmit`s require an entry in the dictionary " *
+                "`emission_limit`. For $p there is none."
+            )
+        end
+    end
+
+    for p ∈ keys(emission_limit(modeltype))
+        em_limit = emission_limit(modeltype, p)
+        # Check for the strategic periods
+        if isa(em_limit, StrategicProfile) && check_timeprofiles
+            @assert_or_log(
+                length(em_limit.vals) == length(𝒯ᴵⁿᵛ),
+                "The timeprofile provided for resource `" * string(p) * "` in the field " *
+                "`emission_limit` does not match the strategic structure."
+            )
+        end
+
+        # Check for potential indexing problems
+        message = "are not allowed for the resource: " * string(p) * " in the Dictionary " *
+            "`emission_limit`."
+        check_strategic_profile(em_limit, message)
+    end
+
+    for p ∈ keys(emission_price(modeltype))
+        em_limit = emission_price(modeltype, p)
+        # Check for the strategic periods
+        if isa(em_limit, StrategicProfile) && check_timeprofiles
+            @assert_or_log(
+                length(em_limit.vals) == length(𝒯ᴵⁿᵛ),
+                "The timeprofile provided for resource `" * string(p) * "` in the field " *
+                "`emission_price` does not match the strategic structure."
+            )
+        end
+
+        # Check for potential indexing problems
+        message = "are not allowed for the resource: " * string(p) * " in the Dictionary " *
+            "`emission_price`."
+        check_strategic_profile(em_limit, message)
+    end
+end
 
 """
     check_time_structure(n::Node, 𝒯)
@@ -145,15 +261,20 @@ function check_profile(fieldname, value::StrategicProfile, 𝒯::TwoLevel)
     len_vals = length(value.vals)
     len_simp = length(𝒯ᴵⁿᵛ)
     if len_vals > len_simp
-        message = "' is shorter than the strategic time structure. \
-        Its last values $(len_vals - len_simp) will be omitted."
+        message = "' is longer than the strategic time structure. \
+        Its last $(len_vals - len_simp) value(s) will be omitted."
     elseif len_vals < len_simp
         message = "' is shorter than the strategic time structure. It will use the last \
-        value for the last $(len_simp - len_vals) strategic periods."
+        value for the last $(len_simp - len_vals) strategic period(s)."
     end
     @assert_or_log len_vals == len_simp "Field '" * string(fieldname) * message
     for t_inv ∈ 𝒯ᴵⁿᵛ
-        check_profile(fieldname, value.vals[t_inv.sp], t_inv.operational, t_inv.sp)
+        check_profile(
+            fieldname,
+            value.vals[minimum([t_inv.sp, length(value.vals)])],
+            t_inv.operational,
+            t_inv.sp,
+        )
     end
 end
 function check_profile(fieldname, value, 𝒯::TwoLevel)
@@ -180,11 +301,11 @@ function check_profile(
     len_simp = length(ts)
     if len_vals > len_simp
         message = "' in strategic period $(sp)  is longer than the operational time  \
-        structure. Its last values $(len_vals - len_simp) will be omitted."
+        structure. Its last $(len_vals - len_simp) value(s) will be omitted."
     elseif len_vals < len_simp
         message = "' in strategic period $(sp) is shorter than the operational \
         time structure. It will use the last value for the last $(len_simp - len_vals) \
-        operational periods."
+        operational period(s)."
     end
     @assert_or_log len_vals == len_simp "Field '" * string(fieldname) * message
 end
@@ -251,11 +372,140 @@ end
 check_profile(fieldname, value, ts, sp) = nothing
 
 """
+    check_strategic_profile(time_profile::TimeProfile, message::String)
+
+Function for checking that an individual `TimeProfile` does not include the wrong type for
+strategic indexing
+
+## Checks
+- `TimeProfile`s access in `StrategicPeriod`s cannot include `OperationalProfile`, \
+`ScenarioProfile`, or `RepresentativeProfile` as this is not allowed through indexing \
+on the `TimeProfile`.
+"""
+function check_strategic_profile(time_profile::TimeProfile, message::String)
+
+    @assert_or_log(
+        !isa(time_profile, OperationalProfile),
+        "Operational profiles " * message
+    )
+    @assert_or_log(
+        !isa(time_profile, ScenarioProfile),
+        "Scenario profiles " * message
+    )
+    @assert_or_log(
+        !isa(time_profile, RepresentativeProfile),
+        "Representative profiles " * message
+    )
+
+    bool_sp = !isa(time_profile, OperationalProfile) *
+        !isa(time_profile, ScenarioProfile) *
+        !isa(time_profile, RepresentativeProfile)
+
+    if isa(time_profile, StrategicProfile)
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:OperationalProfile}),
+            "Operational profiles in strategic profiles " * message
+        )
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:ScenarioProfile}),
+            "Scenario profiles in strategic profiles " * message
+        )
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:RepresentativeProfile}),
+            "Representative profiles in strategic profiles " * message
+        )
+
+        bool_sp *= !isa(time_profile.vals, Vector{<:OperationalProfile}) *
+            !isa(time_profile.vals, Vector{<:ScenarioProfile}) *
+            !isa(time_profile.vals, Vector{<:RepresentativeProfile})
+    end
+
+    return bool_sp
+end
+
+"""
+    check_representative_profile(time_profile::TimeProfile, message::String)
+
+Function for checking that an individual `TimeProfile` does not include the wrong type for
+representative periods indexing
+
+## Input
+- `time_profile` - The time profile that should be checked.
+- `message` - A message that should be printed after the type of profile.
+
+## Checks
+- `TimeProfile`s access in `RepresentativePeriod`s cannot include `OperationalProfile` \
+or `ScenarioProfile` as this is not allowed through indexing on the `TimeProfile`.
+"""
+function check_representative_profile(time_profile::TimeProfile, message::String)
+
+    @assert_or_log(
+        !isa(time_profile, OperationalProfile),
+        "Operational profiles " * message
+    )
+    @assert_or_log(
+        !isa(time_profile, ScenarioProfile),
+        "Scenario profiles " * message
+    )
+
+    bool_rp = !isa(time_profile, OperationalProfile) *
+        !isa(time_profile, ScenarioProfile)
+
+
+    if isa(time_profile, StrategicProfile)
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:OperationalProfile}),
+            "Operational profiles in strategic profiles " * message
+        )
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:ScenarioProfile}),
+            "Scenario profiles in strategic profiles " * message
+        )
+
+        bool_rp *= !isa(time_profile.vals, Vector{<:OperationalProfile}) *
+            !isa(time_profile.vals, Vector{<:ScenarioProfile})
+    end
+
+    return bool_rp
+end
+
+"""
+    check_scenario_profile(time_profile::TimeProfile, message::String)
+
+Function for checking that an individual `TimeProfile` does not include the wrong type for
+scenario indexing
+
+## Checks
+- `TimeProfile`s access in `RepresentativePeriod`s cannot include `OperationalProfile` \
+or `ScenarioProfile` as this is not allowed through indexing on the `TimeProfile`.
+"""
+function check_scenario_profile(time_profile::TimeProfile, message::String)
+
+    @assert_or_log(
+        !isa(time_profile, OperationalProfile),
+        "Operational profiles " * message
+    )
+
+    bool_scp = !isa(time_profile, OperationalProfile)
+
+    if isa(time_profile, StrategicProfile)
+        @assert_or_log(
+            !isa(time_profile.vals, Vector{<:OperationalProfile}),
+            "Operational profiles in strategic profiles " * message
+        )
+
+        bool_scp *= !isa(time_profile.vals, Vector{<:OperationalProfile})
+    end
+
+    return bool_scp
+end
+
+"""
     check_node(n::Node, 𝒯, modeltype::EnergyModel)
 
 Check that the fields of a `Node` corresponds to required structure.
 """
-function check_node(n::Node, 𝒯, modeltype::EnergyModel)
+function check_node(n::Node, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 end
 """
     check_node(n::Availability, 𝒯, modeltype::EnergyModel)
@@ -263,7 +513,7 @@ end
 This method checks that an `Availability` node is valid. By default, that does not include
 any checks.
 """
-function check_node(n::Availability, 𝒯, modeltype::EnergyModel)
+function check_node(n::Availability, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 end
 """
     check_node(n::Source, 𝒯, modeltype::EnergyModel)
@@ -277,9 +527,11 @@ node or that a new `Source` type receives a new method for `check_node`.
 ## Checks
  - The field `cap` is required to be non-negative.
  - The values of the dictionary `output` are required to be non-negative.
- - The value of the field `fixed_opex` is required to be non-negative.
+ - The value of the field `fixed_opex` is required to be non-negative and
+   accessible through a `StrategicPeriod` as outlined in the function
+   `check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)`.
 """
-function check_node(n::Source, 𝒯, modeltype::EnergyModel)
+function check_node(n::Source, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
 
@@ -288,13 +540,10 @@ function check_node(n::Source, 𝒯, modeltype::EnergyModel)
         "The capacity must be non-negative."
     )
     @assert_or_log(
-        sum(opex_fixed(n, t_inv) ≥ 0 for t_inv ∈ 𝒯ᴵⁿᵛ) == length(𝒯ᴵⁿᵛ),
-        "The fixed OPEX must be non-negative."
-    )
-    @assert_or_log(
         sum(outputs(n, p) ≥ 0 for p ∈ outputs(n)) == length(outputs(n)),
         "The values for the Dictionary `output` must be non-negative."
     )
+    check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)
 end
 """
     check_node(n::NetworkNode, 𝒯, modeltype::EnergyModel)
@@ -309,9 +558,11 @@ important that a new `NetworkNode` type includes at least the same fields as in 
  - The field `cap` is required to be non-negative.
  - The values of the dictionary `input` are required to be non-negative.
  - The values of the dictionary `output` are required to be non-negative.
- - The value of the field `fixed_opex` is required to be non-negative.
+ - The value of the field `fixed_opex` is required to be non-negative and
+   accessible through a `StrategicPeriod` as outlined in the function
+   `check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)`.
 """
-function check_node(n::NetworkNode, 𝒯, modeltype::EnergyModel)
+function check_node(n::NetworkNode, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
 
@@ -327,10 +578,7 @@ function check_node(n::NetworkNode, 𝒯, modeltype::EnergyModel)
         sum(outputs(n, p) ≥ 0 for p ∈ outputs(n)) == length(outputs(n)),
         "The values for the Dictionary `output` must be non-negative."
     )
-    @assert_or_log(
-        sum(opex_fixed(n, t_inv) ≥ 0 for t_inv ∈ 𝒯ᴵⁿᵛ) == length(𝒯ᴵⁿᵛ),
-        "The fixed OPEX must be non-negative."
-    )
+    check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)
 end
 """
     check_node(n::Storage, 𝒯, modeltype::EnergyModel)
@@ -346,9 +594,11 @@ important that a new `Storage` type includes at least the same fields as in the
  - The value of the field `stor_cap` is required to be non-negative.
  - The values of the dictionary `input` are required to be non-negative.
  - The values of the dictionary `output` are required to be non-negative.
- - The value of the field `fixed_opex` is required to be non-negative.
+ - The value of the field `fixed_opex` is required to be non-negative and
+   accessible through a `StrategicPeriod` as outlined in the function
+   `check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)`.
 """
-function check_node(n::Storage, 𝒯, modeltype::EnergyModel)
+function check_node(n::Storage, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
     cap = capacity(n)
@@ -369,10 +619,7 @@ function check_node(n::Storage, 𝒯, modeltype::EnergyModel)
         sum(outputs(n, p) ≥ 0 for p ∈ outputs(n)) == length(outputs(n)),
         "The values for the Dictionary `output` must be non-negative."
     )
-    @assert_or_log(
-        sum(opex_fixed(n, t_inv) ≥ 0 for t_inv ∈ 𝒯ᴵⁿᵛ) == length(𝒯ᴵⁿᵛ),
-        "The fixed OPEX must be non-negative."
-    )
+    check_fixed_opex(n, 𝒯ᴵⁿᵛ, check_timeprofiles)
 end
 """
     check_node(n::Sink, 𝒯, modeltype::EnergyModel)
@@ -390,7 +637,7 @@ or that a new `Source` type receives a new method for `check_node`.
  - The sum of the values `:deficit` and `:surplus` in the dictionary `penalty` has to be \
  non-negative to avoid an infeasible model.
 """
-function check_node(n::Sink, 𝒯, modeltype::EnergyModel)
+function check_node(n::Sink, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
     @assert_or_log(
         sum(capacity(n, t) ≥ 0 for t ∈ 𝒯) == length(𝒯),
         "The capacity must be non-negative."
@@ -412,20 +659,53 @@ function check_node(n::Sink, 𝒯, modeltype::EnergyModel)
         )
     end
 end
+"""
+    check_fixed_opex(n::Node, 𝒯ᴵⁿᵛ, check_timeprofiles::Bool)
 
+Checks that the fixed opex value follows the given TimeStructure.
+
+## Checks
+- The `opex_fixed` time profile cannot have a finer granulation than `StrategicProfile`.
+
+## Conditional checks (if `check_timeprofiles=true`)
+- The profiles in `opex_fixed` have to have the same length as the number of strategic
+  periods.
+"""
+function check_fixed_opex(n::Node, 𝒯ᴵⁿᵛ, check_timeprofiles::Bool)
+
+    if isa(opex_fixed(n), StrategicProfile) && check_timeprofiles
+        @assert_or_log(
+            length(opex_fixed(n).vals) == length(𝒯ᴵⁿᵛ),
+            "The timeprofile provided for the field `opex_fixed` does not match the " *
+            "strategic structure."
+        )
+    end
+
+    # Check for potential indexing problems
+    message = "are not allowed for the field `opex_fixed`."
+    bool_sp = check_strategic_profile(opex_fixed(n), message)
+
+    # Check that the value is positive in all cases
+    if bool_sp
+        @assert_or_log(
+            sum(opex_fixed(n, t_inv) ≥ 0 for t_inv ∈ 𝒯ᴵⁿᵛ) == length(𝒯ᴵⁿᵛ),
+            "The fixed OPEX must be non-negative."
+        )
+    end
+end
 
 """
-    check_node_data(n::Node, data::Data, 𝒯, modeltype::EnergyModel)
+    check_node_data(n::Node, data::Data, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
 Check that the included `Data` types of a `Node` corresponds to required structure.
 This function will always result in a multiple error message, if several instances of the
 same supertype is loaded.
 """
-check_node_data(n::Node, data::Data, 𝒯, modeltype::EnergyModel) = nothing
+check_node_data(n::Node, data::Data, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool) = nothing
 
 
 """
-    check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel)
+    check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
 Check that the included `Data` types of a `Node` corresponds to required structure.
 This function will always result in a multiple error message, if several instances of the
@@ -437,7 +717,7 @@ same supertype is loaded.
 - The value of the field `co2_capture` is required to be in the range ``[0, 1]``, if \
 [`CaptureData`](@ref) is used.
 """
-function check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel)
+function check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
     em_data = filter(data -> typeof(data) <: EmissionsData, node_data(n))
     @assert_or_log(
@@ -450,11 +730,12 @@ function check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyMo
 
     for p ∈ process_emissions(data)
         value = process_emissions(data, p)
+        !check_timeprofiles && continue
         !isa(value, TimeProfile) && continue
         check_profile(string(p)*" process emissions", value, 𝒯)
     end
 end
-function check_node_data(n::Node, data::CaptureData, 𝒯, modeltype::EnergyModel)
+function check_node_data(n::Node, data::CaptureData, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
 
     em_data = filter(data -> typeof(data) <: EmissionsData, node_data(n))
     @assert_or_log(
@@ -464,6 +745,7 @@ function check_node_data(n::Node, data::CaptureData, 𝒯, modeltype::EnergyMode
 
     for p ∈ process_emissions(data)
         value = process_emissions(data, p)
+        !check_timeprofiles && continue
         !isa(value, TimeProfile) && continue
         check_profile(string(p)*" process emissions", value, 𝒯)
     end