+++ test/test_checks.jl
@@ -1,6 +1,179 @@
 # Set the global to true to suppress the error message
 EMB.TEST_ENV = true
 
+@testset "Test checks - case dictionary" begin
+    # Resources used in the analysis
+    Power = ResourceCarrier("Power", 0.0)
+    CO2 = ResourceEmit("CO2", 1.0)
+
+    # Function for setting up the system
+    function simple_graph()
+
+        resources = [Power, CO2]
+        ops = SimpleTimes(5, 2)
+        T = TwoLevel(2, 2, ops; op_per_strat=10)
+
+        source = RefSource(
+            "source_emit",
+            FixedProfile(4),
+            FixedProfile(0),
+            FixedProfile(10),
+            Dict(Power => 1),
+        )
+        sink = RefSink(
+            "sink",
+            FixedProfile(3),
+            Dict(:surplus => FixedProfile(-4), :deficit => FixedProfile(4)),
+            Dict(Power => 1),
+        )
+
+        ops = SimpleTimes(5, 2)
+        T = TwoLevel(2, 2, ops; op_per_strat=10)
+
+        nodes = [source, sink]
+        links = [Direct(12, source, sink)]
+        model = OperationalModel(
+            Dict(CO2 => FixedProfile(100)),
+            Dict(CO2 => FixedProfile(0)),
+            CO2
+        )
+        case = Dict(
+                    :T => T,
+                    :nodes => nodes,
+                    :links => links,
+                    :products => resources,
+        )
+        return case, model
+    end
+
+    # Check that the keys are present
+    # - EMB.check_case_data(case)
+    case, model = simple_graph()
+    for key ∈ keys(case)
+        case_test = deepcopy(case)
+        pop!(case_test, key)
+        @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+    end
+
+    # Check that the keys are of the correct format and do not include any unwanted types
+    # - EMB.check_case_data(case)
+    case_test = deepcopy(case)
+    case_test[:T] = 10
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+    case_test = deepcopy(case)
+    case_test[:nodes] = [case[:nodes], case[:nodes], 10]
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+    case_test = deepcopy(case)
+    case_test[:links] = [case[:links], 10]
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+    case_test = deepcopy(case)
+    case_test[:products] = [case[:products], case[:products], 10]
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+end
+
+@testset "Test checks - modeltype" begin
+    # Resources used in the analysis
+    Power = ResourceCarrier("Power", 0.0)
+    CO2 = ResourceEmit("CO2", 1.0)
+
+    # Function for setting up the system
+    function simple_graph()
+
+        resources = [Power, CO2]
+        ops = SimpleTimes(5, 2)
+        T = TwoLevel(2, 2, ops; op_per_strat=10)
+
+        source = RefSource(
+            "source_emit",
+            FixedProfile(4),
+            FixedProfile(0),
+            FixedProfile(10),
+            Dict(Power => 1),
+        )
+        sink = RefSink(
+            "sink",
+            FixedProfile(3),
+            Dict(:surplus => FixedProfile(-4), :deficit => FixedProfile(4)),
+            Dict(Power => 1),
+        )
+
+        ops = SimpleTimes(5, 2)
+        T = TwoLevel(2, 2, ops; op_per_strat=10)
+
+        nodes = [source, sink]
+        links = [Direct(12, source, sink)]
+        case = Dict(
+                    :T => T,
+                    :nodes => nodes,
+                    :links => links,
+                    :products => resources,
+        )
+        return case
+    end
+
+    # Create a function for running the simple graph
+    function run_simple_graph(co2_limit::TS.TimeProfile)
+        case = simple_graph()
+        model = OperationalModel(
+            Dict(CO2 => co2_limit),
+            Dict(CO2 => FixedProfile(0)),
+            CO2
+        )
+        return create_model(case, model), case, model
+    end
+
+    # Extract the default values
+    case = simple_graph()
+    model = OperationalModel(
+        Dict(CO2 => FixedProfile(100)),
+        Dict(CO2 => FixedProfile(0)),
+        CO2
+    )
+
+    # Check that all resources present in the case data are included in the emission limit
+    # - EMB.check_model(case, modeltype::EnergyModel, check_timeprofiles)
+    case_test = deepcopy(case)
+    case_test[:products] = [Power, CO2, ResourceEmit("NG", 0.06)]
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+
+    # Check that the timeprofiles for emission limit and price are correct
+    # - EMB.check_model(case, modeltype::EnergyModel, check_timeprofiles)
+    model = OperationalModel(
+        Dict(CO2 => StrategicProfile([100])),
+        Dict(CO2 => FixedProfile(0)),
+        CO2
+    )
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+    model = OperationalModel(
+        Dict(CO2 => FixedProfile(0)),
+        Dict(CO2 => StrategicProfile([100])),
+        CO2
+    )
+    @test_throws AssertionError run_model(case_test, model, HiGHS.Optimizer)
+
+    # Check that we receive an error if the profiles are wrong
+    # - check_strategic_profile(time_profile, message)
+    rprofile = RepresentativeProfile([FixedProfile(100)])
+    scprofile = ScenarioProfile([FixedProfile(100)])
+    oprofile = OperationalProfile(ones(100))
+
+    co2_limit = oprofile
+    @test_throws AssertionError run_simple_graph(co2_limit)
+    co2_limit = scprofile
+    @test_throws AssertionError run_simple_graph(co2_limit)
+    co2_limit = rprofile
+    @test_throws AssertionError run_simple_graph(co2_limit)
+    co2_limit = StrategicProfile([4])
+    @test_throws AssertionError run_simple_graph(co2_limit)
+
+    co2_limit = StrategicProfile([oprofile, oprofile, oprofile, oprofile])
+    @test_throws AssertionError run_simple_graph(co2_limit)
+    co2_limit = StrategicProfile([scprofile, scprofile, scprofile, scprofile])
+    @test_throws AssertionError run_simple_graph(co2_limit)
+    co2_limit = StrategicProfile([rprofile, rprofile, rprofile, rprofile])
+    @test_throws AssertionError run_simple_graph(co2_limit)
+end
+
 @testset "Test checks - emission data" begin
     # Resources used in the analysis
     Power = ResourceCarrier("Power", 0.0)
@@ -44,21 +217,42 @@ EMB.TEST_ENV = true
                     :links => links,
                     :products => resources,
         )
-        return run_model(case, model, HiGHS.Optimizer), case, model
+        return case, model
+    end
+
+    # Create a function for running the simple graph
+    function run_simple_graph(em_data::Vector{<:EmissionsData})
+        case, model = simple_graph(em_data)
+        return create_model(case, model), case, model
     end
 
     # Test that only a single EmissionData is allowed
     # - EMB.check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel)
     em_data = [EmissionsProcess(Dict(CO2 => 10.0)), EmissionsEnergy()]
-    @test_throws AssertionError simple_graph(em_data)
+    @test_throws AssertionError run_simple_graph(em_data)
 
     # Test that the capture rate is bound by 0 and 1
     # - EMB.check_node_data(n::Node, data::CaptureData, 𝒯, modeltype::EnergyModel)
     em_data = [CaptureEnergyEmissions(1.2)]
-    @test_throws AssertionError simple_graph(em_data)
-
+    @test_throws AssertionError run_simple_graph(em_data)
     em_data = [CaptureEnergyEmissions(-1.2)]
-    @test_throws AssertionError simple_graph(em_data)
+    @test_throws AssertionError run_simple_graph(em_data)
+
+    # Test that the timeprofile check is working and only showing the warning once
+    # - EMB.check_node_data(n::Node, data::EmissionsData, 𝒯, modeltype::EnergyModel
+    em_data = [EmissionsProcess(Dict(CO2 => StrategicProfile([1])))]
+    @test_throws AssertionError run_simple_graph(em_data)
+    case, model = simple_graph(em_data)
+    msg = "Checking of the time profiles is deactivated:\n" *
+    "Deactivating the checks for the time profiles is strongly discouraged. " *
+    "While the model will still run, unexpected results can occur, as well as " *
+    "inconsistent case data.\n\n" *
+    "Deactivating the checks for the timeprofiles should only be considered, " *
+    "when testing new components. In all other instances, it is recommended to " *
+    "provide the correct timeprofiles using a preprocessing routine.\n\n" *
+    "If timeprofiles are not checked, inconsistencies can occur."
+    @test_logs (:warn, msg) create_model(case, model; check_timeprofiles=false)
+    @test_logs (:warn, msg) for k ∈ [1,2] create_model(case, model; check_timeprofiles=false) end
 
 end
 
@@ -106,7 +300,13 @@ end
                     :links => links,
                     :products => resources,
         )
-        return run_model(case, model, HiGHS.Optimizer), case, model
+        return case, model
+    end
+
+    # Create a function for running the simple graph
+    function run_simple_graph(T::TimeStructure, tp::TimeProfile)
+        case, model = simple_graph(T, tp)
+        return create_model(case, model), case, model
     end
 
     day = SimpleTimes(24, 1)
@@ -116,15 +316,15 @@ end
     ts = TwoLevel(2, 1, day)
     ops = OperationalProfile(ones(24))
     tp = StrategicProfile([ops, ops, ops])
-    @test_throws AssertionError simple_graph(ts, tp)
+    @test_throws AssertionError run_simple_graph(ts, tp)
 
     # Test that there is an error with wrong operational profiles
     # - EMB.check_profile(fieldname, value::OperationalProfile, ts::SimpleTimes, sp)
     ts = TwoLevel(2, 1, day)
     tp = OperationalProfile(ones(20))
-    @test_throws AssertionError simple_graph(ts, tp)
+    @test_throws AssertionError run_simple_graph(ts, tp)
     tp = OperationalProfile(ones(30))
-    @test_throws AssertionError simple_graph(ts, tp)
+    @test_throws AssertionError run_simple_graph(ts, tp)
 
     # Test that there is warning when using OperationalProfile with RepresentativePeriods
     # - EMB.check_profile(fieldname, value::RepresentativeProfile, ts::SimpleTimes, sp)
@@ -135,7 +335,7 @@ end
     It only works reasonable if all representative periods have an operational \
     time structure of the same length. Otherwise, the last value is repeated. \
     The system is tested for the all representative periods."
-    @test_logs (:warn, msg) simple_graph(ts, tp)
+    @test_logs (:warn, msg) run_simple_graph(ts, tp)
 
     # Test that there is warning when using RepresentativeProfile without RepresentativePeriods
     # - EMB.check_profile(fieldname, value::RepresentativeProfile, ts::SimpleTimes, sp)
@@ -143,13 +343,25 @@ end
     tp = RepresentativeProfile([FixedProfile(5), FixedProfile(10)])
     msg = "Field cap: Using `RepresentativeProfile` with `SimpleTimes` is dangerous, as it \
     may lead to unexpected behaviour. In this case, only the first profile is used and tested."
-    @test_logs (:warn, msg) simple_graph(ts, tp)
+    @test_logs (:warn, msg) run_simple_graph(ts, tp)
 
     # Test that there is an error when `RepresentativeProfile` have a different length than
     # the corresponding `RepresentativePeriods`
     # - EMB.check_profile(fieldname, value::RepresentativeProfile, ts::SimpleTimes, sp)
     ts = TwoLevel(2, 1, RepresentativePeriods(3, 8760, ones(3)/3, [day, day, day]))
-    @test_throws AssertionError simple_graph(ts, tp)
+    @test_throws AssertionError run_simple_graph(ts, tp)
+
+    # Check that turning of the timeprofile checks leads to a warning
+    case, model = simple_graph(ts, tp)
+    msg = "Checking of the time profiles is deactivated:\n" *
+    "Deactivating the checks for the time profiles is strongly discouraged. " *
+    "While the model will still run, unexpected results can occur, as well as " *
+    "inconsistent case data.\n\n" *
+    "Deactivating the checks for the timeprofiles should only be considered, " *
+    "when testing new components. In all other instances, it is recommended to " *
+    "provide the correct timeprofiles using a preprocessing routine.\n\n" *
+    "If timeprofiles are not checked, inconsistencies can occur."
+    @test_logs (:warn, msg) create_model(case, model; check_timeprofiles=false)
 end
 
 @testset "Test checks - Nodes" begin
@@ -179,7 +391,7 @@ end
                     :links => links,
                     :products => resources,
         )
-        return run_model(case, model, HiGHS.Optimizer), case, model
+        return create_model(case, model), case, model
     end
 
     # Test that the fields of a Source are correctly checked
@@ -214,14 +426,25 @@ end
         @test_throws AssertionError simple_graph(source, sink)
 
         # Test that a wrong fixed OPEX is caught by the checks.
-        source = RefSource(
-            "source",
-            FixedProfile(4),
-            FixedProfile(10),
-            FixedProfile(-5),
-            Dict(Power => 1),
-        )
-        @test_throws AssertionError simple_graph(source, sink)
+        function check_opex_prof(opex_fixed, sink)
+            source = RefSource(
+                "source",
+                FixedProfile(4),
+                FixedProfile(10),
+                opex_fixed,
+                Dict(Power => 1),
+            )
+            return simple_graph(source, sink)
+        end
+        opex_fixed = FixedProfile(-5)
+        @test_throws AssertionError check_opex_prof(opex_fixed, sink)
+
+        # Test that a wrong profile for fixed OPEX is caught by the checks.
+        # - check_fixed_opex(n::Node, 𝒯ᴵⁿᵛ, check_timeprofiles::Bool)
+        opex_fixed = StrategicProfile([1])
+        @test_throws AssertionError check_opex_prof(opex_fixed, sink)
+        opex_fixed = OperationalProfile([1])
+        @test_throws AssertionError check_opex_prof(opex_fixed, sink)
 
         # Test that correct input solves the model to optimality.
         source = RefSource(
@@ -231,7 +454,8 @@ end
             FixedProfile(5),
             Dict(Power => 1),
         )
-        m , _, _ = simple_graph(source, sink)
+        _, case, model = simple_graph(source, sink)
+        m = run_model(case, model, HiGHS.Optimizer)
         @test termination_status(m) == MOI.OPTIMAL
     end
 
@@ -282,7 +506,8 @@ end
             Dict(:surplus => FixedProfile(4), :deficit => FixedProfile(10)),
             Dict(Power => 1),
         )
-        m , _, _ = simple_graph(source, sink)
+        _, case, model = simple_graph(source, sink)
+        m = run_model(case, model, HiGHS.Optimizer)
         @test termination_status(m) == MOI.OPTIMAL
     end
 
@@ -326,7 +551,7 @@ end
                     :links => links,
                     :products => resources,
         )
-        return run_model(case, model, HiGHS.Optimizer), case, model
+        return create_model(case, model), case, model
     end
 
     # Test that the fields of a NetworkNode are correctly checked
@@ -386,7 +611,8 @@ end
             Dict(NG => 2),
             Dict(Power => 1),
         )
-        m , _, _ = simple_graph(network)
+        _, case, model = simple_graph(network)
+        m = run_model(case, model, HiGHS.Optimizer)
         @test termination_status(m) == MOI.OPTIMAL
     end
 
@@ -438,7 +664,7 @@ end
                     :links => links,
                     :products => resources,
         )
-        return run_model(case, model, HiGHS.Optimizer), case, model
+        return create_model(case, model), case, model
     end
 
     # Test that the fields of a Storage are correctly checked
@@ -532,7 +758,8 @@ end
             Dict(Power => 1, aux => 0.05),
             Dict(Power => 1),
         )
-        m , _, _ = simple_graph(storage)
+        _, case, model = simple_graph(storage)
+        m = run_model(case, model, HiGHS.Optimizer)
         @test termination_status(m) == MOI.OPTIMAL
     end
 end