+++ NEWS.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## Version 0.6.7 (2024-03-21)
+
+* Allow for deactivation of timeprofile checks while printing a warning in this case.
+* Fixed a bug for a too short `StrategicProfile` in the checks.
+* Added checks for the case dictionary.
+* Extended checks for the modeltype
+* Added functions that can be used to check whether a `TimeProfile` can be indexed over `StrategicPeriod`s, `RepresentativePeriod`s, or `OperationalScenario`s.
+
 ## Version 0.6.6 (2024-03-04)
 
 ### Examples